Introducing Null object pattern
