+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/dynapi/NullValidator.java
@@ -1,0 +1,9 @@
+package edu.cornell.mannlib.vitro.webapp.dynapi;
+
+public class NullValidator implements ModelValidator {
+
+    public boolean isValid(String uri){
+        return true;
+    }
+
+}
